Usuniecie Testowych Klas, Poprawa Aktywatora kart. Uporzadkowanie Folderow.
