+++ .gitignore
@@ -77,3 +77,4 @@ ExportedObj/
 /Unity/Unity/Library
 /Unity/Unity/Logs
 /Unity/Unity/Temp
+/Unity/.vsconfig
