+++ Unity/Assets/Resources/cards_img/card1.png.meta
