+++ Unity/Assets/Resources/cards_img/card2.png.meta
