+++ Unity/Assets/Scenes/Board.unity
@@ -553,7 +553,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 527042593}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c2147a7b7502b5b4eb79ee9696d9bece, type: 3}
+  m_Script: {fileID: 11500000, guid: 787339ae4258d274b847f8bad61cc407, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &721407938
@@ -1092,7 +1092,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   maxSize: 8
-  minSize: 0
   cards: []
 --- !u!1 &1714867615
 GameObject:
