+++ Unity/Assets/ScriptableObjects.meta
