+++ Unity/Assets/ScriptableObjects/Karta1.asset.meta
