+++ Unity/Assets/ScriptableObjects/Karta2.asset.meta
