+++ Unity/Assets/Scripts/CardActivator.cs.meta
