+++ Unity/Assets/Scripts/FightManager.cs
@@ -2,8 +2,16 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class display : MonoBehaviour
+public class FightManager : MonoBehaviour
 {
+    private Stack stack;
+    private Player player;
+
+    public FightManager(Stack stack, Player player) 
+    {
+        this.stack = stack;
+        this.player = player;
+    }
     // Start is called before the first frame update
     void Start()
     {
