--- Unity/Assets/Scripts/drag.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ee07d18d430424d4c85dc5e6b5b96dab
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
