--- Unity/Assets/Scripts/move.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e307a6ccd182641499922875822612a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
