--- Unity/Assets/StarterCards.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c327126e47108664186f2dce4a0b9cdf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
