--- Unity/Assets/Tossed.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8c168e6e12dc29c44a925d2b6286a99b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
