--- Unity/Assets/display.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c2147a7b7502b5b4eb79ee9696d9bece
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
